Moved StreamControl/Loader to plugins
