+++ frictionless/loaders/__init__.py
@@ -1,2 +1,1 @@
-from .stream import StreamLoader
 from .text import TextLoader
