+++ tests/plugins/test_stream.py
